requirements.txt
# 常见依赖清单
ccxt==4.2.14          # 核心交易库
pandas==2.2.3         # 数据处理（与 pandas-ta-remake 版本对齐）
pandas-ta-remake==1.0.4  # 技术指标库 (ATR, RSI)，兼容 import pandas_ta
python-dotenv==1.0.0  # 环境变数管理
loguru==0.7.2         # 极简且强大的日志库
pyyaml==6.0.1         # 配置文件解析
aiohttp==3.9.1        # 异步网络请求